imagenet/requirements.txt
numpy~=1.21.1
gym~=0.18.0
tqdm~=4.58.0
stable-baselines3~=1.1.0
einops~=0.3.0
wandb~=0.11.0
accelerate~=0.4.0
pandas~=1.2.5
tensorboard~=2.5.0
scipy~=1.7.0
pillow~=8.2.0
torchattacks~=3.0.0
torchmetrics~=0.5.0
easydict